Added files so analysis can begin
The binned_data is the preliminary B0 candidates produced by Vid's ML code. It is then binned in ranges of q^2 in the same way as the angular observables are defined over specific ranges of q^2.
The predictions are the standard model observables in a JSON file. This is copied from the skeleton code folder and code will be supplied to load in these predictions.