quick analysis using much of skeleton code
Using the binned data from preliminary ML cut to produce fits. Much of the code is heavily copied from the skeleton code. The Standard Model prediction is also plotted for a qualitative analysis.
Requires binned_data and predictions folder downloaded
Also needs iminuit to be installed in python for the minimisation - not natively installed in anaconda